Usability improvements to new generic tracking data arc
